clocky: a port of my c2-demo to webfui
Clocky is evidently my hello-world program when getting
used to monitored atoms and svg drawing. This program
simply introduces a watched atom via launch-app
like other good webfui demos. The differences
though are that:

 1) the output of this watch function is svg
 2) the atom itself is being banged on to provide animation

Point (1) is fairly straightfoward, but involved me pulling
in c2 as a dependency for now since I wanted to just use
the svg arc function and math helpers there.

Point (2) is achieved via a call to requestAnimationFrame()
for smooth and efficient rendering in browsers that support
it. Note that there is a polyfill javascript for older
browsers which fallsback to a simple timer based animation.
There are links in rAF.js to more information about this.

Some final salient points while I am thinking about them.
This example is almost exactly like the c2-demo, with the
exception that the render-all callback takes a state
argument instead of assuming have that function reach
outside itself to grab the relevant atom. While that
seems a trivial matter, it did cascade into me changing
the higher logic to emphasize that the atom itself could
be local to the update function and the watch instigated
by launch-app. However, since the requestAnimationFrame
callback itself assumes no arguments, the clock-atom is
captured in the closure of a function which is called
immediately, which seems just a little awkard, but probably
worth the overall locality of variables.